Attribute sets: fix logic for replacing bold/italic in CJK languages
At some point in the past, the italic.replacement.color went from being
a background color to being a text color. While a transparent background
makes some sense, transparent text is never readable.

Also, even when using XEP, in Western languages, we do not want all bold
text to be gray => removed that.

Added 80,80,80 to color list... would have been great to
avoid that. &dark-gray-old; seems close, but it is not pure gray, it has
some color mixed in which is rather undesirable.