Be more consistent in avoiding page breaks after :
When the final character of a para is : , we try to avoid a page break.
This behavior has been in the stylesheets for some time, but it so far
did not work in e.g. procedures, because para/step has its own template
that so far did not care about our page break shenanigans