Screen trimmer: allow removing start and end whitespace in screens
The previous implementation of the trim-whitespace-in-screens
functionality had the issue that it could only cut away whitespace from
either the beginning or the end.

If a screen had whitespace both at the beginning and at the end, only
the beginning of the screen would be trimmed. That should now be fixed