suse2013 fo: fix procedures in qandasets (re: suse-xsl #200)
This fixes part of issue #200, the part where we don't show procedures
inside questions.

Test qandaset:

  <qandaset>
    <qandaentry>
     <question>
      <para>
       How to do?
      </para>
     </question>
     <answer>
      <procedure>
       <step>
        <para>Just do.</para>
       </step>
      </procedure>
     </answer>
    </qandaentry>
  </qandaset>

Adding a preamble <para/> at the beginning of the procedure fixed this.
Adding a para before the procedure also fixed this.

Why exactly we need to remove the preamble fo:block that is basically
empty to get our procedure to display, I don't know. But it works.

The (positive) upshot is that this also fixes the weird space at the top
of procedures without an introductory para.