Increase footer cutoff factor a little for CJK langs
This avoids adding an ellipsis to some longish section title in the
footer, i.e. this footer:

325            My incredibly long section titl...  MyCoolProduct 1.0

might become:

325       My incredibly long section title, yeah!  MyCoolProduct 1.0

(Now imagine the same examples with CJK text.)

Let's be a bit less careful here. However, it should be clear that the
footer implementation is still overdue for a makeover. It currently does
not take into account the length of the (abbreviated) product name. As
long as that does not happen, there is little use in trying to
accurately predict the right point at which to add the ellipsis in the
footer