Avoid xsltproc's infinite recursion error in long screens
The CAP release notes have a screen with a length in excess of 1000
lines. This would trigger xsltproc to complain about an infinite
recursion in the stylesheets, as the "splitscreen" template that is
supposed to check line length in screens would recurse until it had
gone through all lines of the screen.

We now do two things differently:
* We stop the search for long lines in a screen after finding the first
  one
* After 250 lines in a single screen we print a warning.