SUSE2013 CSS: End the negative margin insanity
It does feel good to close so many awful home-made margin issues at
once. Thank you, ::before. (IE 10 only knows :before, not the standard
::before, but at this point I am hoping nobody is using it anymore.)

This fix is not perfect (e.g. does not work properly for figures),
but much better than what we had before.

This kills the "scroll-to-ID" animation we used to have, leaving only
the animation for the "Top" (of page) button intact. It's not
super-compatible with the :target::before code and I did not want to
investigate too much more. Since this whole functionality was broken for
the past 2+ years anyway, I am guessing nobody will mind.

This also fixes :target highlighting in parts that were broken much
before.

Cf. https://www.bryanbraun.com/anchorjs/#pages-with-a-sticky-navbar