Fix #422: Add "last modified" line on cover page
This is a proof-of-concept:

* Introduce parameter "last.modified" for FO and XHTML
  by default it's empty
* Introduce a "last.modified" key in language file
  (currently en ATM)
* When the cover page is created, only add "Last modified"
  line when $last.modified != ''.

Currently, the parameter $last.modified is used "as-is"; it
will be included in the respective line no matter if it's a
date or not.
However, if it's a date, it is currently inconsistent with the
publication date. That date is created from this line:

<date><?dbtimestamp format="B d, Y"?></date>

In other words, to make both dates to be consistent, the
last modified date needs to be transformed into the format
of the PI inside <date>.