suse2021/fo: initial color, logo updates
Font update needs to be done later, Poppins unfortunately is available
in too many weights for FOP and we get the 100 weight instead of 400
... again, just like with Work Sans before. In addition, Poppins is
only available for Latin, whereas Open Sans is also available for
Greek (granted, irrelevant) and Cyrillic (relevant). Meaning, we will
need to stick to Open Sans (or maybe use Roboto, which is a little more
in spirit with Poppins, though not much).