suse2022/sass/css: Move mobile menu button to the right
The mobile menu button used to interfere with the document title a bit
much. It's a bit illogical to click a button on the right that makes a
menu appear on the left, but it's not _that_ bad.
(The new placement also makes it even more apparent that the button
design/placement is mostly nicked from fandom.com, but so be it.)