Fix #499: Change format of <title>
For SEO reasons, the format needs to be changed.
This was the old format:

  <CURRENT_TITLE> | <GUIDE_TITLE> | <PRODUCT> <VERSION>

As <PRODUCT> can be quite long (for example, "SUSE Linux Enterprise
Server..."), we use the abbreviated product name. This is taken from
d:productname[@role="abbrev"]. The new format looks like this now:

  <PRODUCT_ABBREV> <VERSION>: <CURRENT_TITLE> | <GUIDE_TITLE>

(We don't shorten the title to make sure we can distinguish the page)

See also DOCTEAM-781