Fix #533: Let publication date appear after titles
This is only useful for SmartDocs and other articles

Previously, the publication date was placed _after_
the abstract. However, if the abstract becomes long,
the publication date is moved. Perhaps it is even moved
ouside of the visiable area of the reader.

This fix improves readability so readers now right after
reading the title when this article was published.