Implement Qualtrics rating
From the meeting with Asya, we need an empty
<div> container with an id.

For this to work, two parameter are introduced:

* generate.qualtrics.div (default 0)
  When this parameter is set to 1, the empty div
  container is created.

* qualtrics.id (default "qualtrics_container")
  The default id to reference in Qualtrics. Any
  non-empty value will be used.
  This parameter should normally not changed.

Keep in mind, only when $generate.qualtrics.div != 0
AND $qualtrics.id != '' are set that the empty <div> is
generated. This avoids a <div id=""> container.

Additionally, Qualtrics requires some JavaScript code
in the header.