Separate Qualtrics JS script
Due to XML parsing errors ("<" and "&"), we had to separate
the JavaScript into a separate file. A xsl:copy-of or xsl:value-of
of the direct JS code doesn't work and leads to these XML errors. :-(

Masking the two characters as &lt; and &amp; would lead to literal &lt;
or &amp; strings in the code. This doesn't work either.

Parameter qualtrics-feedback.js points to static/js/qualtrics-feedback.js

Co-authored-by: Gayathri Gandaboyina <gayathri.gandaboyina@suse.com>