Fix Adjust <link rel="next|prev">
This commit preserves the prev/next link in the content, but
fixes the prev/next link in the <head>.

When chunking HTML, the upstream DocBook stylesheets generate
links which points to the HTML pages that comes before or after.
This looks like this:

  <link rel="prev" href="..." />
  <link rel="next" href="..." />

When setting $rootid to point to a specific article or book, the
next and prev links have to be only in the boundary of this realm.

Pointing next to the following article or book leads to broken links
in Siteimprove. Same applies for prev and the previous article or book.

This fix adjust the creation of <link rel="..."> links.

Related: DOCTEAM-1772, #dad3c438