Fix language for DC-*all deliverables
Use `(ancestor-or-self::*[@xml:lang][1])/@xml:lang` to extract
the language of the nearest current element.

This means, if a chapter is the current element and has a xml:lang
attribute, it is used. If it doesn't, the XPath expression tries
any of its ancestors. If the parent has the attribute, it is used,
otherwise it can go up til the root element is reached.